I have finished all studio

+++ src/studio0/BurritoBudget.java
@@ -6,9 +6,9 @@ public class BurritoBudget {
 		int burritosEatenWeek1 = 7;
 		int burritosEatenWeek2 = 4;
 		int burritosEatenWeek3 = 19; //don't judge me
-		int burritosEatenWeek4 = 1;
+		int burritosEatenWeek4 = 1;	
 		
-		int totalBurritos = burritosEatenWeek1 + burritosEatenWeek2 + burritosEatenWeek3 + burritosEatenWeek4;
+		double totalBurritos = burritosEatenWeek1 + burritosEatenWeek2 + burritosEatenWeek3 + burritosEatenWeek4;
 		double averageWeeklyBurritos = totalBurritos / 4;
 		
 		double burritoPrice = 7.50;
